+++ src/gui-news.c
@@ -317,7 +317,7 @@ static XmxCallback (include_fsb_cb)
   FILE *fp;
   char line[MO_LINE_LENGTH], *status;
 
-  mo_window *win = mo_fetch_window_by_id (XmxExtractUniqid ((int)client_data));
+  mo_window *win = mo_fetch_window_by_id(XmxExtractUniqid(client_data));
 
   if (!win)
     return;
@@ -383,7 +383,7 @@ static XmxCallback (include_fsb_cb)
 
 static XmxCallback (include_button_cb) /* Why is this here ?*/
 {
-  mo_window *win = mo_fetch_window_by_id (XmxExtractUniqid ((int)client_data));
+  mo_window *win = mo_fetch_window_by_id(XmxExtractUniqid(client_data));
   return;
 }
 
@@ -396,10 +396,10 @@ char *NNTPgetquoteline(char *art);
 
 static XmxCallback (news_win_cb)
 {
-  mo_window *win = mo_fetch_window_by_id (XmxExtractUniqid ((int)client_data));
+  mo_window *win = mo_fetch_window_by_id(XmxExtractUniqid(client_data));
   char *msg,*subj,*group,*from,*line;
 
-  switch (XmxExtractToken ((int)client_data))
+  switch (XmxExtractToken(client_data))
     {
     case 0: /* POST */
       XtUnmanageChild (win->news_win);
@@ -461,11 +461,11 @@ static XmxCallback (news_win_cb)
 
 static XmxCallback (follow_win_cb)
 {
-  mo_window *win = mo_fetch_window_by_id (XmxExtractUniqid ((int)client_data));
+  mo_window *win = mo_fetch_window_by_id(XmxExtractUniqid(client_data));
   char *msg,*subj,*group,*from,*line;
   int pos;
 
-  switch (XmxExtractToken ((int)client_data))
+  switch (XmxExtractToken(client_data))
     {
     case 0: /* POST */
       XtUnmanageChild (win->news_win);
